+++ src/_help-text.js
@@ -1,0 +1,18 @@
+export default `
+Usage:
+  tap-arc <options>
+
+Parses TAP data from stdin, and outputs a "spec-like" formatted result.
+
+Options:
+
+  -v | --verbose
+    Output full stack trace
+
+  -p | --pessimistic | --bail
+    Immediately exit upon encountering a failure
+    example: tap-arc -p
+
+  --no-color
+    Output without ANSI escape sequences for colors
+    example: tap-arc --no-color`
