+++ test/create-upstream-error-tap.cjs
@@ -1,9 +1,9 @@
 const test = require('tape')
 
 // eslint-disable-next-line import/no-unresolved, no-unused-vars
-const { hello } = require('./this-is-a-fake-path')
+// const { hello } = require('./this-is-a-fake-path')
 
-test('tests pass, but this should exit code 0', (t) => {
+test('these tests pass, but this should exit code 0', (t) => {
   t.ok('OKAY')
   t.end()
 })
